cache product list to mitigate slow response from api
todd believes the issue with slow response on product lists has something to do with the default sort order (Name, ID). When the catalog gets sufficiently large the sorting algorithm’s performance decrements substantially. He’s looking into indexing around some of that but this will be a fix for now